bounce buffer removed and tested

+++ filesys/cache.h
@@ -5,8 +5,8 @@
 #include "devices/block.h"
 
 void cache_init(void);
-void cache_read(struct block*, block_sector_t, void*, off_t, int);
-void cache_write(struct block*, block_sector_t, void*, off_t, int);
+void cache_read(struct block*, block_sector_t, void*, unsigned, int);
+void cache_write(struct block*, block_sector_t, void*, unsigned, int);
 void cache_flush(void);
 
 #endif /* filesys/cache.h */
